// ViedeoFrontend/src/app/swagger/model/languageDto.ts
/**
 * Diplomarbeit
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LanguageDto { 
    id?: number;
    countryName?: string | null;
}