// ViedeoFrontend/src/app/swagger/model/register.ts
/**
 * Diplomarbeit
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Register { 
    mail?: string | null;
    userName?: string | null;
    password?: string | null;
    countryName?: string | null;
    firstName?: string | null;
    lastName?: string | null;
}